Add navigation and dummy screens
The structure for the screens folder was chosen to be able to add styles
file and others and export everything from all the files related to a
component while also only exporting the main Component through the
screens module. Although none of the current screens export types or
style objects, it prepares for a more expandable structure.